Combine export logic for OneDrive and SharePoint
